+++ ReleaseNotes.md
@@ -1,5 +1,8 @@
 # Release Notes
 KsWare.Presentation.Interfaces
 
+#### 2019-09-14 (v0.1)
+- NEW KsWare.Presentation.Interfaces.Plugins.DataTemplateConverter
+
 #### 2019-09-13 (v0.1)
 - NEW